Add bench crate
This is the crate I used to inform #4's performance optimizations
(such as function inlining). I used it in combination with both a
performance profiler (`perf`) and `cargo bench`.

Included with the crate is documentation on how to use it for
benchmarking and performance profiling, and guidance on how this
information is used for `getargs`' development.

Namely, it establishes:

- `getargs` utilizes performance-guided optimization in its development

- `getargs` is currently the fastest argument parsing crate I know of,
  and PRs will be accepted that add new benched crates that outperform
  `getargs` (or not)

- There is a focus on making sure `getargs` continues to be fast, and
  performance regressions will be reviewed on a case-by-base basis, but
  being fast is not the only goal (i.e. it is secondary)

- The `bench` crate is being published so that performance testing
  and benchmarking is easy, results can be easily verified, anyone can
  contribute, etc. it's not just "performance tests LoganDark can run"